Use traceback module to specify CODE_{LINE,FUNC,FILE}
$python3 -c 'import journald; x = lambda: journald.send("tescik"); x()'

gives a message like:

MESSAGE=tescik
CODE_FILE=<string>
CODE_LINE=1
CODE_FUNC=<lambda>
_TRANSPORT=journal
...

This makes the CODE_* fields much more useful: before they would
always give the location of the sd_journald_sendv() call in the module
source, and now they specify the location of the caller of
journald.send().

When using the low-level journald.sendv() is becomes the
responsibility of the caller to specify the CODE_* fields.