journal: fix compatibility with python2
This is a lazy workaround: 4c9a241949067fc8d55f3ea12170ad364bd8b18d
is amended to do nothing on python2, so we have the same issue that
was present before. This allows the code to execute, and hopefully
almost nobody is using python2 code anyway.

f868a56b935b6152d611b22f7a5538f14dafb194 is amended in the same way.
For python2 code we have the same lack of timezone-awareness as before.

This allows the tests to pass under python 2.7.