Add a long_description to setup.py
Currently the `systemd-python` package on PyPI renders with a "Project description" of
"The author of this package has not provided a project description" because the setup.py
has no `long_description` argument.

As a result, the rendering of the project on PyPi appears either immature or not legitimate.

Ths adds in a `long_description` argument which displays the `README.md` as the "Project description"
on PyPI.