reader: drop kwargs declaration in the _reader.Journal.add_match registration
Drop specification of kwargs in the _reader.Journal.add_match registration. The
handling was never implemented, so any any kwargs would be summarilly ignored.

The handling of kwargs is done in the python wrapper, so no kwargs are passed
to the native C function (and would be ignored if they were). IIRC, the plan
was initially to do this in the C extension, but then we realized that this is
very much not a hot path and doing in the the wrapper is just fine.