meson: add test to run pytest
The python binary that was initially found is used to launch pytest.

It turns out to be surprisingly hard to run 'pytest' in test() with the python
binary detected using meson's python module. test() does not allow the
"program" argument to be a str, which is what python module's .full_path()
returns. So let's "find" the program again. This results in one line of noise
in meson output, but I don't see a nicer way.

'ninja -C build test' works. 'meson test -C build' is better, because it can be
called with '-v' to print details of the test.